add ListAdapter to MainActivity

+++ app/src/main/java/com/nandohidayat/app/inputdatabarang/Barang.java
@@ -47,7 +47,7 @@ public class Barang {
         HargaJual = hargaJual;
     }
 
-    public double getStok() {
+    public int getStok() {
         return Stok;
     }
